refactor database
This commit refactors the database into 3 tables, with the eventual goal of allowing changing the names of things by keying each item based on ID instead of name. All pages and queries have been changed to reflect the new database structure. In addition, this has made it very easy to add autocomplete to the name insert, so I've done that as well. You can press esc to autofill based on the name displayed in grey. This helps with matching items that have multiple barcodes. Finally, this commit changes the input field on the main page into a textarea instead, so multi-row input is possible. I'll have to see how it displays on the raspberry pi when my cache finally decides to update... It will likely see css changes for portrait mode, where it definitely doesn't work on some browsers.