fix broken multi-barcode
The latest commit broke items that have the same name but more than one barcode associated with them. This commit fixes that problem, though I'm not terribly happy with the fix... It involves an extra query on top of potentially 3 inserts, making `/add_barcode_and_item` by far the heaviest endpoint in terms of performance. I don't really anticipate it being a problem, but I hope a more elegant solution is on the horizon.