items clickable
This commit ensures that all items are now consistently clickable to get to their respective item pages. This adds a lot of UI consistency to the application.
With all items being clickable, the add/remove workflow is now much more practical. I have considered changing the "HOME" button on the item page to "back" to return to the previous page instead, and I think it's likely I'll end up doing that...