renaming item_id
By convention, I had refactored 'item_id' to just 'id' in responses, since all other id's are internal to the database. I missed add_barcode, and that has now been rectified. This had previously caused an error where adding an existing item via barcode and then attempting to click on it in the left pane before refreshing would cause an issue, but now it's fixed. I'm using the web interface for this commit since I'm too lazy to switch to linux or download and setup the project on windows.